Migrate Query Source page to React: Query Editor section (top-right)
