+++ client/app/components/queries/QueryEditor/index.less
@@ -1,36 +1,13 @@
-.editor__wrapper {
-  padding: 15px;
-  margin-bottom: 10px;
-  height: 100%;
-  display: flex;
-  flex-direction: column;
-  flex-wrap: nowrap;
+.query-editor-container {
+  margin-bottom: 0;
+  position: relative;
 
-  .editor__container {
-    margin-bottom: 0;
-    flex: 1 1 auto;
-    position: relative;
-
-    .ace_editor {
-      position: absolute;
-      left: 0;
-      top: 0;
-      width: 100%;
-      height: 100%;
-      margin: 0;
-    }
-
-    &[data-executing] {
-      .ace_marker-layer {
-        .ace_selection {
-          background-color: rgb(255, 210, 181);
-        }
-      }
-    }
-  }
-
-  .query-editor-controls {
-    flex: 0 0 auto;
-    margin-top: 10px;
+  .ace_editor {
+    position: absolute;
+    left: 0;
+    top: 0;
+    width: 100%;
+    height: 100%;
+    margin: 0;
   }
 }
